RunSummary refactor
